requirements.txt
flask
flask-cors
requests
gunicorn
python-telegram-bot==20.7
apscheduler
